feat: add canvas/sprite size and new file commands along with an expression parser utility
