+++ src/app/commands/cmd_canvas_size.cpp
@@ -32,207 +32,7 @@
 #include <memory>
 
 #include "app/console.h"
-#include <cmath>
-#include <cctype>
-#include <stdexcept>
-#include <algorithm>
-
-class ExprParser {
-  std::string src;
-  std::size_t pos;
-  double baseValue;
-  bool hasError;
-  std::string errorMsg;
-
-  void skipWhitespace() {
-    while (pos < src.size() && std::isspace(static_cast<unsigned char>(src[pos]))) {
-      pos++;
-    }
-  }
-
-  bool matchStr(const std::string& s) {
-    if (pos + s.size() > src.size()) return false;
-    for (std::size_t i = 0; i < s.size(); ++i) {
-      if (std::tolower(static_cast<unsigned char>(src[pos + i])) != std::tolower(static_cast<unsigned char>(s[i]))) {
-        return false;
-      }
-    }
-    pos += s.size();
-    return true;
-  }
-
-public:
-  ExprParser(const std::string& input, double baseVal)
-      : src(input), pos(0), baseValue(baseVal), hasError(false) {
-    skipWhitespace();
-    if (pos < src.size()) {
-      char c = src[pos];
-      if (c == '+' || c == '-' || c == '*' || c == '/') {
-        src = std::to_string(baseValue) + src.substr(pos);
-        pos = 0;
-      }
-    }
-  }
-
-  double parseExpression() {
-    double val = parseTerm();
-    while (true) {
-      skipWhitespace();
-      if (pos >= src.size()) break;
-      char op = src[pos];
-      if (op == '+' || op == '-') {
-        pos++;
-        double rhs = parseTerm();
-        if (op == '+') val += rhs;
-        else val -= rhs;
-      } else {
-        break;
-      }
-    }
-    return val;
-  }
-
-  double parseTerm() {
-    double val = parseFactor();
-    while (true) {
-      skipWhitespace();
-      if (pos >= src.size()) break;
-      char op = src[pos];
-      if (op == '*' || op == '/') {
-        pos++;
-        double rhs = parseFactor();
-        if (op == '*') {
-          val *= rhs;
-        } else {
-          if (rhs == 0.0) {
-            hasError = true;
-            errorMsg = "Division by zero";
-            val = 0.0;
-          } else {
-            val /= rhs;
-          }
-        }
-      } else {
-        break;
-      }
-    }
-    return val;
-  }
-
-  double parseFactor() {
-    skipWhitespace();
-    if (pos >= src.size()) {
-      hasError = true;
-      errorMsg = "Unexpected end of expression";
-      return 0.0;
-    }
-
-    bool negative = false;
-    if (src[pos] == '+') {
-      pos++;
-      skipWhitespace();
-    } else if (src[pos] == '-') {
-      pos++;
-      negative = true;
-      skipWhitespace();
-    }
-
-    double val = 0.0;
-    if (src[pos] == '(') {
-      pos++;
-      val = parseExpression();
-      skipWhitespace();
-      if (pos < src.size() && src[pos] == ')') {
-        pos++;
-      } else {
-        hasError = true;
-        errorMsg = "Missing closing parenthesis";
-      }
-    } else if (std::isdigit(static_cast<unsigned char>(src[pos])) || src[pos] == '.') {
-      std::size_t nextPos = 0;
-      try {
-        val = std::stod(src.substr(pos), &nextPos);
-        pos += nextPos;
-      } catch (...) {
-        hasError = true;
-        errorMsg = "Invalid number format";
-      }
-    } else {
-      hasError = true;
-      errorMsg = std::string("Unexpected character: '") + src[pos] + "'";
-      pos++;
-    }
-
-    if (negative) {
-      val = -val;
-    }
-
-    while (true) {
-      skipWhitespace();
-      if (pos >= src.size()) break;
-      if (src[pos] == '%') {
-        val *= 0.01;
-        pos++;
-      } else if (matchStr("px")) {
-        continue;
-      } else {
-        break;
-      }
-    }
-
-    return val;
-  }
-
-  bool getResult(double& outResult) {
-    try {
-      outResult = parseExpression();
-      skipWhitespace();
-      if (pos < src.size()) {
-        if (matchStr("px")) {
-          skipWhitespace();
-        }
-        if (pos < src.size()) {
-          hasError = true;
-          errorMsg = std::string("Unexpected trailing characters: '") + src.substr(pos) + "'";
-        }
-      }
-    } catch (const std::exception& e) {
-      hasError = true;
-      errorMsg = e.what();
-    }
-    return !hasError;
-  }
-
-  std::string getError() const { return errorMsg; }
-};
-
-// Evaluate an arithmetic expression relative to a base value.
-// Supports binary expressions (e.g. 16*5, 100+50, 200/2) and
-// prefix operators (+N, -N, *N, /N — relative to base).
-// Trailing suffixes like "px" or "%" are stripped automatically.
-static int evalExpr(const std::string& rawText, int base, bool reportError = false) {
-  double result = 0.0;
-  ExprParser parser(rawText, base);
-  if (parser.getResult(result)) {
-    return (int)std::round(result);
-  } else {
-    if (reportError) {
-      app::Console console;
-      console.printf("Error evaluating expression '%s': %s\n", rawText.c_str(), parser.getError().c_str());
-    }
-    return base;
-  }
-}
-
-// Returns true if 'rawText' contains an arithmetic operator (binary or prefix).
-static bool isExpr(const std::string& rawText) {
-  for (char c : rawText) {
-    if (c == '+' || c == '-' || c == '*' || c == '/' || c == '%' || c == '(' || c == ')' || c == '.') {
-      return true;
-    }
-  }
-  return false;
-}
+#include "app/util/expr_parser.h"
 
 namespace app {
